+++ dms-core/src/main/kotlin/team/aliens/dms/common/service/feign/FeignService.kt
@@ -1,0 +1,8 @@
+package team.aliens.dms.common.service.feign
+
+import team.aliens.dms.common.annotation.Service
+
+@Service
+class FeignService(
+    neisFeignService: NeisFeignService
+) : NeisFeignService by neisFeignService
