+++ dms-core/src/main/kotlin/team/aliens/dms/common/service/feign/NeisFeignService.kt
@@ -1,0 +1,10 @@
+package team.aliens.dms.common.service.feign
+
+import team.aliens.dms.domain.meal.dto.NeisMealServiceDietInfosResponse
+import team.aliens.dms.domain.school.dto.NeisSchoolResponse
+
+interface NeisFeignService {
+    fun getNeisSchoolInfo(schoolName: String, schoolAddress: String): NeisSchoolResponse
+
+    fun getNeisMealServiceDietInfo(sdSchoolCode: String, regionCode: String): NeisMealServiceDietInfosResponse
+}
